Clean up account creation, remove some old methods
