+++ tests-manual/account/creation
@@ -1,0 +1,15 @@
+#!/usr/bin/env python3
+from ocflib.account.creation import create_account
+
+
+if __name__ == '__main__':
+    create_account(
+        user='ggroup2',
+        password='hunter2',
+        real_name='Gee Group Two',
+        email='ckuehl+ggroup2@ocf.berkeley.edu',
+        calnet_uid=123123123,
+        callink_oid=None,
+        keytab='/home/c/ck/ckuehl/create.keytab',
+        admin_principal='create/admin',
+    )
